requirements.txt
Flask==2.3.3
flask-cors==4.0.0
selenium==4.15.0
webdriver-manager==4.0.1
Pillow
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
google-genai
google-api-core